Remove extraneous file
